refactor scripting into calculate navbar menu
