+++ baby-gru/src/components/navbar-menus/MoorhenEditMenu.js
@@ -19,7 +19,6 @@ export const MoorhenEditMenu = (props) => {
             <MoorhenAddRemoveHydrogenAtomsMenuItem key='add_remove_hydrogens' {...menuItemProps}/>
             <MoorhenMergeMoleculesMenuItem key="merge" {...menuItemProps} />
             <MoorhenDeleteUsingCidMenuItem key="delete" {...menuItemProps} />
-            <MoorhenSuperposeMenuItem key="superpose_structures" {...menuItemProps} />
             <MoorhenCopyFragmentUsingCidMenuItem key="copy_fragment" {...menuItemProps} />
             <MoorhenGoToMenuItem key="go_to_cid" {...menuItemProps} />
             {props.extraEditMenuItems && props.extraEditMenuItems.map( menu => menu)}
